faceverification/requirements.txt
numpy
opencv-python
pytorch
facenet-pytorch